refactor: segments 폴더 추가
