+++ src/templates/Team/segments/TeamMaxPeopleInput.tsx
